[refactor] : 코드 리포멧팅
