+++ core/src/main/java/dev/handsup/user/dto/response/UserJoinResponse.java
@@ -3,4 +3,5 @@ package dev.handsup.user.dto.response;
 public record UserJoinResponse(
 
 	Long userId
-) {}
+) {
+}
