+++ core/src/testFixtures/java/dev/handsup/fixture/AuctionFixture.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = PRIVATE)
 public class AuctionFixture {
 
-	public static Auction auction(){
+	public static Auction auction() {
 		return Auction.of(
 			"거의 새상품 버즈 팔아요",
 			ProductCategory.of("디지털 기기"),
